+++ src/main/java/lotto/domain/YieldCalculation.java
@@ -5,7 +5,7 @@ import lotto.input.LottoTickets;
 public class YieldCalculation {
     private static final int PERCENT = 100;
 
-    public double revenue(int totalPrizePool) {
+    public double revenue(long totalPrizePool) {
         LottoTickets lottoTickets = new LottoTickets();
         int paid = lottoTickets.getPrice();
         return (((double) totalPrizePool / paid) * PERCENT);
